Re-enable bootstrapping tests
